requirements.txt
flask==3.0.3
google-genai==1.36.0
google-generativeai==0.8.5
langchain-core==0.3.29
langchain-google-genai==2.0.10
langgraph==0.6.7
python-dotenv==1.0.1
gunicorn==21.2.0